requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
gradio>=3.50.0
rembg>=2.0.0
onnxruntime>=1.15.0